+++ src/components/content/observations/patient-observations.tsx
@@ -5,12 +5,13 @@ import { usePatientBuilderDiagnosticReports } from "@/fhir/diagnostic-report";
 
 export type PatientObservationsProps = {
   className?: cx.Argument;
+  enableFQS?: boolean;
 };
 
-const Component = ({ className }: PatientObservationsProps) => {
-  const diagnosticReports = usePatientBuilderDiagnosticReports();
+const Component = ({ className, enableFQS = false }: PatientObservationsProps) => {
+  const diagnosticReports = usePatientBuilderDiagnosticReports(enableFQS);
 
-  return <PatientObservationsBase className={className} query={diagnosticReports} />;
+  return <PatientObservationsBase className={className} query={diagnosticReports} enableFQS />;
 };
 
 export const PatientObservations = withErrorBoundary(Component, "PatientObservations");