+++ src/components/content/timeline/patient-timeline.tsx
@@ -1,35 +1,45 @@
 import cx from "classnames";
 import { useRef } from "react";
+import { patientTimelineColumns } from "./helpers/columns";
 import { defaultTimelineFilters, timelineFilters } from "./helpers/filters";
 import { usePatientEncounterDetailsDrawer } from "./helpers/modal-hooks";
 import { defaultTimelineSort, timelineSortOptions } from "./helpers/sorts";
-import { patientTimelineColumns } from "./patient-timeline-columns";
 import { useObservationsDetailsDrawer } from "../observations/helpers/drawer";
+import { getDateRangeView } from "../resource/helpers/view-date-range";
 import { ResourceTableActions } from "../resource/resource-table-actions";
 import { ResourceTable } from "@/components/content/resource/resource-table";
 import { useCTW } from "@/components/core/providers/use-ctw";
+import { TimelineEventModel } from "@/fhir/models/timeline-event";
 import { useTimelineEvents } from "@/fhir/timeline-event";
 import { useFilteredSortedData } from "@/hooks/use-filtered-sorted-data";
 
 export type PatientTimelineProps = {
   className?: cx.Argument;
+  enableFQS?: boolean;
 };
 
-export function PatientTimeline({ className }: PatientTimelineProps) {
-  const timelineEventsQuery = useTimelineEvents();
+export function PatientTimeline({ className, enableFQS = false }: PatientTimelineProps) {
+  const timelineEventsQuery = useTimelineEvents(enableFQS);
   const containerRef = useRef<HTMLDivElement>(null);
   const { featureFlags } = useCTW();
-  const { data, setFilters, setSort } = useFilteredSortedData({
+  const { viewOptions, defaultView } = getDateRangeView<TimelineEventModel>("date");
+  const { data, setFilters, setSort, setViewOption } = useFilteredSortedData({
+    defaultView,
     defaultFilters: defaultTimelineFilters,
     defaultSort: defaultTimelineSort,
     records: timelineEventsQuery.data,
   });
   const openEncounterDetails = usePatientEncounterDetailsDrawer();
-  const openDiagnosticReportDetails = useObservationsDetailsDrawer();
+  const openDiagnosticReportDetails = useObservationsDetailsDrawer(enableFQS);
 
   return (
     <div className={cx(className, "ctw-scrollable-pass-through-height")} ref={containerRef}>
       <ResourceTableActions
+        viewOptions={{
+          onChange: setViewOption,
+          options: viewOptions,
+          defaultView,
+        }}
         filterOptions={{
           onChange: setFilters,
           defaultState: defaultTimelineFilters,