+++ src/fhir/action-helper.ts
@@ -1,6 +1,6 @@
 import { FhirResource, Resource } from "fhir/r4";
 import { v4 as uuidv4 } from "uuid";
-import { omitEmptyArrays } from "./client";
+import { fixupFHIR } from "./client";
 import { isFhirError } from "./errors";
 import { ASSEMBLER_CODING, CREATE_CODING, createProvenance } from "./provenance";
 import { CTWRequestContext } from "@/components/core/providers/ctw-context";
@@ -12,7 +12,7 @@ export async function createOrEditFhirResource(
   resource: Resource,
   requestContext: CTWRequestContext
 ) {
-  const { fhirClient } = requestContext;
+  const { fhirClient, fhirWriteBackClient } = requestContext;
   const resourceModified = resource;
   let action = "create";
   let response;
@@ -20,18 +20,20 @@ export async function createOrEditFhirResource(
   try {
     if (resource.id) {
       action = "update";
+      // Use the ODS client to update a resource
       response = await fhirClient.update({
         resourceType: resource.resourceType,
         id: resource.id,
-        body: omitEmptyArrays(resource),
+        body: fixupFHIR(resource),
       });
       if (!isFhirError(response)) {
         await createProvenance("UPDATE", response, requestContext);
       }
     } else {
-      response = await fhirClient.create({
+      // Utilize the FHIR write-back client for creating resources
+      response = await fhirWriteBackClient.create({
         resourceType: resource.resourceType,
-        body: omitEmptyArrays(resource),
+        body: fixupFHIR(resource),
       });
       if (!isFhirError(response)) {
         resourceModified.id = response.id;
@@ -101,7 +103,7 @@ export async function createFhirResourceWithProvenance(
   };
 
   return fhirClient.transaction({
-    body: omitEmptyArrays({
+    body: fixupFHIR({
       ...bundle,
       type: "transaction",
     }),