+++ CHANGELOG.md
@@ -1,5 +1,38 @@
 # ctw-component-library
 
+## 1.13.1
+
+### Patch Changes
+
+- fa1fb932: Fix bug when trying to save or update resources fetched from FQS.
+
+## 1.13.0
+
+### Minor Changes
+
+- 4ac4c517: Add view filtering by date for timeline, documents, and outside medication components.
+
+## 1.12.0
+
+### Minor Changes
+
+- 57e03269: Added `fhirWriteBackClient` to proxy creation of new FHIR resources through the ehr-data-integration server.
+
+### Patch Changes
+
+- 88c8d13c: Update some dependencies
+- bdc77c6f: Fix case where CCDA viewer crashes
+
+## 1.11.0
+
+### Minor Changes
+
+- de1c2c88: Add feature flag option to use FQS (Zus's GraphQL Fast-Query-Service) for fetching conditions.
+
+### Patch Changes
+
+- b411c5f3: Fix issue with superclusters loading too much data when FQS is enabled
+
 ## 1.10.4
 
 ### Patch Changes